refactor: completedの保管位場所変更
